Move to report service, use publishedReportStore
