+++ src/test/kotlin/com/terraformation/backend/accelerator/db/ReportStoreTest.kt
@@ -14,7 +14,6 @@ import com.terraformation.backend.accelerator.model.ExistingProjectReportConfigM
 import com.terraformation.backend.accelerator.model.NewProjectReportConfigModel
 import com.terraformation.backend.accelerator.model.ProjectIndicatorModel
 import com.terraformation.backend.accelerator.model.ProjectIndicatorTargetsModel
-import com.terraformation.backend.accelerator.model.PublishedReportComparedProps
 import com.terraformation.backend.accelerator.model.ReportAutoCalculatedIndicatorEntryModel
 import com.terraformation.backend.accelerator.model.ReportAutoCalculatedIndicatorModel
 import com.terraformation.backend.accelerator.model.ReportAutoCalculatedIndicatorTargetModel
@@ -7630,283 +7629,6 @@ class ReportStoreTest : DatabaseTest(), RunsAsDatabaseUser {
       assertEquals(
           emptyList<AutoCalculatedIndicatorTargetsModel>(),
           store.fetchAutoCalculatedIndicatorTargets(projectId),
-      )
-    }
-  }
-
-  @Nested
-  inner class UnpublishedProperties {
-    private lateinit var reportId: ReportId
-
-    @BeforeEach
-    fun setUp() {
-      insertProjectReportConfig()
-      reportId =
-          insertReport(
-              highlights = "highlights",
-              additionalComments = "additional comments",
-              financialSummaries = "financial summaries",
-          )
-    }
-
-    @Test
-    fun `returns empty list when report has never been published`() {
-      val model = store.fetchOne(reportId)
-      assertEquals(
-          emptyList<PublishedReportComparedProps>(),
-          model.unpublishedProperties,
-          "unpublishedProperties should be empty",
-      )
-    }
-
-    @Test
-    fun `returns empty list when report matches published qualitative properties`() {
-      insertPublishedReport(
-          highlights = "highlights",
-          additionalComments = "additional comments",
-          financialSummaries = "financial summaries",
-      )
-      insertReportAchievement(position = 0, achievement = "Achievement A")
-      insertPublishedReportAchievement(position = 0, achievement = "Achievement A")
-      insertReportChallenge(position = 0, challenge = "Challenge A", mitigationPlan = "Plan A")
-      insertPublishedReportChallenge(
-          position = 0,
-          challenge = "Challenge A",
-          mitigationPlan = "Plan A",
-      )
-
-      assertEquals(
-          emptyList<PublishedReportComparedProps>(),
-          store.fetchOne(reportId).unpublishedProperties,
-          "unpublishedProperties should be empty when nothing changed",
-      )
-    }
-
-    @Test
-    fun `identifies all changed properties`() {
-      val commonIndicatorId = insertCommonIndicator(isPublishable = true)
-      val projectIndicatorId = insertProjectIndicator(isPublishable = true)
-
-      // Published version: old values for all properties
-      insertPublishedReport(
-          highlights = "old highlights",
-          additionalComments = "old comments",
-          financialSummaries = "old financial summaries",
-      )
-      // Published has no achievements or challenges
-      insertPublishedReportCommonIndicator(
-          indicatorId = commonIndicatorId,
-          value = BigDecimal(10),
-      )
-      insertPublishedReportProjectIndicator(
-          indicatorId = projectIndicatorId,
-          value = BigDecimal(30),
-      )
-      insertPublishedReportAutoCalculatedIndicator(
-          indicator = AutoCalculatedIndicator.Seedlings,
-          value = BigDecimal(50),
-      )
-
-      // Current draft: different values for all properties
-      insertReportAchievement(position = 0, achievement = "New Achievement")
-      insertReportChallenge(
-          position = 0,
-          challenge = "New Challenge",
-          mitigationPlan = "New Plan",
-      )
-      insertReportCommonIndicator(indicatorId = commonIndicatorId, value = BigDecimal(20))
-      insertReportProjectIndicator(indicatorId = projectIndicatorId, value = BigDecimal(40))
-      insertReportAutoCalculatedIndicator(
-          indicator = AutoCalculatedIndicator.Seedlings,
-          systemValue = BigDecimal(40),
-          overrideValue = BigDecimal(60),
-      )
-
-      assertEquals(
-          setOf(
-              PublishedReportComparedProps.Highlights,
-              PublishedReportComparedProps.Achievements,
-              PublishedReportComparedProps.Challenges,
-              PublishedReportComparedProps.AdditionalComments,
-              PublishedReportComparedProps.FinancialSummaries,
-              PublishedReportComparedProps.CommonIndicators,
-              PublishedReportComparedProps.ProjectIndicators,
-              PublishedReportComparedProps.AutoCalculatedIndicators,
-          ),
-          store.fetchOne(reportId, includeIndicators = true).unpublishedProperties.toSet(),
-          "Unpublished properties should contain all props",
-      )
-    }
-
-    @Test
-    fun `identifies photos as changed when new photo added`() {
-      insertFile()
-      insertPublishedReport(
-          highlights = "highlights",
-          additionalComments = "additional comments",
-          financialSummaries = "financial summaries",
-      )
-      // Current report has photo; published has none
-      insertReportPhoto()
-
-      assertEquals(
-          listOf(PublishedReportComparedProps.Photos),
-          store.fetchOne(reportId).unpublishedProperties,
-          "Adding a photo causes unpublishedProperties to include photos",
-      )
-    }
-
-    @Test
-    fun `identifies photos as changed when photo removed`() {
-      insertFile()
-      insertPublishedReport(
-          highlights = "highlights",
-          additionalComments = "additional comments",
-          financialSummaries = "financial summaries",
-      )
-      // Insert the photo as deleted in the current report to satisfy the FK on
-      // published_report_photos
-      insertReportPhoto(deleted = true)
-      insertPublishedReportPhoto()
-
-      assertEquals(
-          listOf(PublishedReportComparedProps.Photos),
-          store.fetchOne(reportId).unpublishedProperties,
-          "Removing a photo causes unpublishedProperties to include photos",
-      )
-    }
-
-    @Test
-    fun `does not identify commonIndicators as changed for non-publishable indicators`() {
-      insertCommonIndicator(isPublishable = false)
-      insertPublishedReport(
-          highlights = "highlights",
-          additionalComments = "additional comments",
-          financialSummaries = "financial summaries",
-      )
-      // Published has no common indicator row (non-publishable not published)
-      insertReportCommonIndicator(value = BigDecimal(99))
-
-      val model = store.fetchOne(reportId, includeIndicators = true)
-      assertEquals(
-          emptyList<PublishedReportComparedProps>(),
-          model.unpublishedProperties,
-          "CommonIndicators should not be flagged for non-publishable indicators",
-      )
-    }
-
-    @Test
-    fun `does not identify projectIndicators as changed for non-publishable indicators`() {
-      insertProjectIndicator(isPublishable = false)
-      insertPublishedReport(
-          highlights = "highlights",
-          additionalComments = "additional comments",
-          financialSummaries = "financial summaries",
-      )
-      // Published has no project indicator row (non-publishable not published)
-      insertReportProjectIndicator(value = BigDecimal(99))
-
-      val model = store.fetchOne(reportId, includeIndicators = true)
-      assertEquals(
-          emptyList<PublishedReportComparedProps>(),
-          model.unpublishedProperties,
-          "ProjectIndicators should not be flagged for non-publishable indicators",
-      )
-    }
-
-    @Test
-    fun `uses system value when no override for autoCalculatedIndicators comparison`() {
-      insertPublishedReport(
-          highlights = "highlights",
-          additionalComments = "additional comments",
-          financialSummaries = "financial summaries",
-      )
-      // Publish Seedlings=40 only; other indicators have no published rows and no current rows.
-      insertPublishedReportAutoCalculatedIndicator(
-          indicator = AutoCalculatedIndicator.Seedlings,
-          value = BigDecimal(40),
-      )
-      // System value 40, no override — should match the published value of 40
-      insertReportAutoCalculatedIndicator(
-          indicator = AutoCalculatedIndicator.Seedlings,
-          systemValue = BigDecimal(40),
-      )
-
-      val model = store.fetchOne(reportId, includeIndicators = true)
-      assertEquals(
-          emptyList<PublishedReportComparedProps>(),
-          model.unpublishedProperties,
-          "AutoCalculatedIndicators should not be flagged when system value matches published",
-      )
-    }
-
-    @Test
-    fun `returns correct unpublishedProperties per report when fetching multiple reports`() {
-      // Report 1 (reportId from setUp): never published → unpublishedProperties == emptyList()
-
-      // Report 2: published and unpublished matches exactly → unpublishedProperties == emptyList()
-      val reportId2 =
-          insertReport(
-              startDate = LocalDate.of(2030, Month.JANUARY, 1),
-              endDate = LocalDate.of(2030, Month.MARCH, 31),
-              highlights = "same highlights",
-          )
-      insertPublishedReport(highlights = "same highlights")
-
-      // Report 3: published but highlights differs → unpublishedProperties contains Highlights
-      val reportId3 =
-          insertReport(
-              startDate = LocalDate.of(2030, Month.APRIL, 1),
-              endDate = LocalDate.of(2030, Month.JUNE, 30),
-              highlights = "new highlights",
-          )
-      insertPublishedReport(highlights = "old highlights")
-
-      val reports =
-          store.fetch(projectId = projectId, includeFuture = true, includeIndicators = true)
-      val byId = reports.associateBy { it.id }
-
-      assertEquals(
-          emptyList<PublishedReportComparedProps>(),
-          byId[reportId]?.unpublishedProperties,
-          "Report 1 (never published): unpublishedProperties should be empty",
-      )
-      assertEquals(
-          emptyList<PublishedReportComparedProps>(),
-          byId[reportId2]?.unpublishedProperties,
-          "Report 2 (published, no changes): unpublishedProperties should be empty",
-      )
-      assertEquals(
-          listOf(PublishedReportComparedProps.Highlights),
-          byId[reportId3]?.unpublishedProperties,
-          "Report 3 (highlights changed): unpublishedProperties should just be Highlights",
-      )
-    }
-
-    @Test
-    fun `does not compare progress notes or projectsComments for commonIndicators`() {
-      insertCommonIndicator(isPublishable = true)
-      insertPublishedReport(
-          highlights = "highlights",
-          additionalComments = "additional comments",
-          financialSummaries = "financial summaries",
-      )
-      insertPublishedReportCommonIndicator(
-          value = BigDecimal(10),
-          progressNotes = "old progress notes",
-          projectsComments = "old projects comments",
-      )
-      insertReportCommonIndicator(
-          value = BigDecimal(10),
-          progressNotes = "new progress notes",
-          projectsComments = "new projects comments",
-      )
-
-      val model = store.fetchOne(reportId, includeIndicators = true)
-      assertEquals(
-          emptyList<PublishedReportComparedProps>(),
-          model.unpublishedProperties,
-          "CommonIndicators should not be flagged when only notes/comments changed",
       )
     }
   }
